trainingdiary/requirements.txt
asgiref==3.3.4
bootstrap4==0.1.0
Django==3.2
django-registration==3.0.1
numpy==1.16.3
pandas==1.1.5
python-dateutil==2.8.0
pytz==2021.1
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
xlrd==1.2.0
